# app/OpenApi/Schemas/Responses/ValidationErrorResponse.php
<?php

namespace App\OpenApi\Schemas\Responses;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: "ValidationErrorResponse",
    properties: [
        new OA\Property(property: "message", type: "string"),
        new OA\Property(
            property: "errors",
            type: "object",
            additionalProperties: new OA\AdditionalProperties(
                type: "array",
                items: new OA\Items(type: "string")
            )
        ),
    ]
)]
class ValidationErrorResponse {}